휴가 로직 수정 유급휴가만 휴가 차감되도록

+++ src/main/java/com/mountain/for_mountain/domain/leave/dto/LeaveCreateRequest.java
@@ -20,7 +20,7 @@ public class LeaveCreateRequest {
     private String endDate;
 
     @NotNull
-    private Integer days;
+    private Double days;
 
     private String reason;
 }
