+++ src/main/java/com/mountain/for_mountain/domain/leave/dto/LeaveResponse.java
@@ -13,7 +13,7 @@ public class LeaveResponse {
     private String leaveType;
     private String startDate;
     private String endDate;
-    private Integer days;
+    private Double days;
     private String reason;
     private String status;
     private String appliedAt;
