+++ src/main/java/com/mountain/for_mountain/domain/leave/dto/LeaveUpdateRequest.java
@@ -17,7 +17,7 @@ public class LeaveUpdateRequest {
     private String endDate;
 
     @NotNull
-    private Integer days;
+    private Double days;
 
     private String reason;
 }
